refactor(WinnerMessageBuilder): 우승자 리스트를 토대로 최종 우승자 메시지를 만드는 WinnerMessageBuilder 클래스 추가
