+++ src/main/java/racinggame/controller/GameConsoleController.java
@@ -1,20 +1,20 @@
-package racinggame.view;
+package racinggame.controller;
 
 import racinggame.environments.GlobalVariables;
 import racinggame.model.Car;
 import racinggame.model.Cars;
 import racinggame.model.RacingGame;
 import racinggame.utils.InputParser;
+import racinggame.utils.WinnerMessageBuilder;
+import racinggame.view.InputView;
+import racinggame.view.OutputView;
 
-public class GameConsole {
+public class GameConsoleController {
     private static final int ILLEGAL_GAME_TURN_COUNT = -1;
     private static final String ASK_PLAYER_NAME_MSG = "경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)";
     private static final String ERROR_INPUT_PLAYER_NAME_MSG = "[ERROR] 자동차의 이름은 1글자 이상 5글자 이하여야 합니다. 이름을 다시 입력해주세요..";
     private static final String ASK_GAME_TURN_COUNT_MSG = "시도할 횟수는 몇회인가요?";
     private static final String ERROR_INPUT_GAME_TURN_COUNT_MSG = "[ERROR] 게임 횟수는 양의 정수여야 합니다. 다시 입력하세요...";
-    private static final String WINNER_MSG_PREFIX = "최종 우승자는 ";
-    private static final String WINNER_MSG_POSTFIX = " 입니다.";
-    private static final String WINNER_MSG_DELIMITER = ",";
 
     private RacingGame racingGame;
     private InputParser inputParser;
@@ -25,7 +25,7 @@ public class GameConsole {
     private Cars playerCars;
     private int gameTurnCnt;
 
-    public GameConsole() {
+    public GameConsoleController() {
         this.racingGame = new RacingGame();
         this.inputParser = new InputParser();
 
@@ -104,32 +104,8 @@ public class GameConsole {
     public void playGame() {
         Cars winnerCars = this.racingGame.play(this.playerCars, this.gameTurnCnt);
 
-        String winnerMsg = buildWinnerMsg(winnerCars);
+        String winnerMsg = WinnerMessageBuilder.build(winnerCars);
 
         this.outputView.write(winnerMsg);
-    }
-
-    private String buildWinnerMsg(Cars winnerCars) {
-        StringBuilder builder = new StringBuilder();
-
-        builder.append(WINNER_MSG_PREFIX);
-        for (int i = 0; i < winnerCars.size(); i++) {
-            Boolean isLastWinner = i == winnerCars.size() - 1;
-
-            builder.append(buildWinnerMsgInternal(winnerCars.get(i), isLastWinner));
-        }
-        builder.append(WINNER_MSG_POSTFIX);
-
-        return builder.toString();
-    }
-
-    private String buildWinnerMsgInternal(Car winnerCar, boolean isLastWinner) {
-        String winnerMsg = winnerCar.getPlayerName();
-
-        if (!isLastWinner) {
-            winnerMsg += WINNER_MSG_DELIMITER;
-        }
-
-        return winnerMsg;
     }
 }
